Backspace when cursor is off right edge of screen should move back two cells
This is extremely counter-intuitive for a terminal like kitty that
allows the cursor to be off screen when drawing a char in the last cell.

However, most other terminals dont allow the cursor to be off the right
edge so it makes sense for them. There is no terminfo property
controlling this behavior as far as I know.

The execrable ncurses actually relies on this behavior of backspace to
position the cursor in some cases, the good lord alone knows why.
So in the interests of compatibility lets special case backsapce when
cursor is off the right edge.

A simple test script to see the behavior:

```py
import subprocess

cols = int(subprocess.check_output('tput cols'.split()).decode().strip())
print(end='\x1b[?7l')  # turn off line wrap
print('a' * cols, end='\x08bc')  # print a upto last cell then backspace and bc
input()
print(end='\x1b[?7h')  # turn on line wrap
print()
```

This should result in a line starting with aaaa and ending with bc.
It does so in most terminals, exceptions being foot and kitty before
this patch.

Note that this also means that backspacing when cursor is off the right
edge of the terminal is broken in all terminals with this weird
behavior. See the docs in screen_backspace() for details. Sigh.

Fixes #8841